Feat: Kafka를 이용한 Booking - Payment - User 연동 서비스 구현

+++ payment-service/src/main/java/com/flight_booking/payment_service/presentation/request/UpdateFareRequestDto.java
@@ -6,7 +6,8 @@ import java.util.UUID;
 
 public record UpdateFareRequestDto(
     @NotNull(message = "Booking ID cannot be null") UUID bookingId,
-    @NotNull(message = "Previous Fare cannot be null") @Positive(message = "Fare must be positive") Integer previousFare,
-    @NotNull(message = "Update Fare cannot be null") @Positive(message = "Fare must be positive") Integer newFare) {
+    @NotNull(message = "Previous Fare cannot be null") @Positive(message = "Fare must be positive") Long previousFare,
+    @NotNull(message = "Update Fare cannot be null") @Positive(message = "Fare must be positive") Long newFare) {
 
 }
+
